// src/types/seo.ts
export default class Seo {
  id?: number;
  title?: string;

  addressCountry?: string;

  addressLocality?: string;
  postalCode?: string;
  streetAddress?: string;
  description?: string;
  date?: Date;
  image?: string;
  logo?: string;
  publisher?: string;
  topics?: string;
  video?: string;
  country?: string;
  url?: string;
  lang?: string;

  author?: string;
  keywords?: string;
  feed?: string;
}